Fix a double-free error with ExternalReference.
The ExternalReferenceParser created an ExternalReference object as a
std::shared_ptr. This object is then passed to a callback via a raw
pointer which then passes it to a CityObject that saves the raw pointer
inside a std::unique_ptr.
The problem with this is that after the parsing is done, the shared_ptr
will go away deleting the ExternalReference object. At this point the
unique_ptr in CityObject points to a destroyed object. And this will
lead to a crash when the CityObject is destroyed and the unique_ptr
tries to delete the already deleted ExternalReference object.
To fix this we no longer use a shared_ptr at all but start with a
unique_ptr in the parser that releases its pointer to the callback
thereby giving up its ownership.