Fix std::string in union.
The setup before this commit violates the C++ standard.
If two classes with non-standard constructors and/or
destructors are in a union their constructors and/or
destructors have to be called explicitly when
switching between the value types.

Using a union is also unsafe as it's illegal to
access the "wrong" of the two members. But the API
doesn't have a way to tell which member is the active.

The new solution doesn't use unions but allows clients
to safely access even the wrong value.